requirements.txt
google-genai
huggingface_hub
python-dotenv
PyPDF2
moviepy
requests
